move utils and config tests to unit folder
